requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
openai>=1.60.0
pydantic>=2.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
